Organise the code a bit
