+++ client/my-sites/plans/jetpack-plans/product-lightbox/index.tsx
@@ -8,6 +8,7 @@ import { useDispatch } from 'react-redux';
 import MultipleChoiceQuestion from 'calypso/components/multiple-choice-question';
 import { recordTracksEvent } from 'calypso/state/analytics/actions/record';
 import { useStoreItemInfoContext } from '../product-store/context/store-item-info-context';
+import { PricingBreakdown } from '../product-store/pricing-breakdown/pricing-breakdown';
 import { ProductStoreBaseProps } from '../product-store/types';
 import getProductIcon from '../product-store/utils/get-product-icon';
 import slugToSelectorProduct from '../slug-to-selector-product';
@@ -16,7 +17,6 @@ import { PRODUCT_OPTIONS, PRODUCT_OPTIONS_HEADER } from './constants';
 import { Icons } from './icons/icons';
 import { Tags } from './icons/tags';
 import PaymentPlan from './payment-plan';
-import { PricingBreakdown } from './pricing-breakdown';
 import ProductDetails from './product-details';
 import { getSlugsOfProductsInludedInBundle } from './utils';
