+++ client/my-sites/plans/jetpack-plans/product-store/types.ts
@@ -108,3 +108,17 @@ export type MoreInfoLinkProps = {
 	onClick?: VoidFunction;
 	isExternal?: boolean;
 };
+
+export type PricingBreakdownProps = {
+	product: SelectorProduct;
+	siteId: number | null;
+	includedProductSlugs: ReadonlyArray< string >;
+	showBreakdownHeading?: boolean;
+};
+
+export type PricingBreakdownItem = {
+	name: React.ReactNode;
+	slug: string;
+	originalPrice: number;
+	renderedPrice: React.ReactNode;
+};
